Fix integer version of glm::log2 for GCC
This pretty much reverts the fix done in commit
1ed0e3865b5de0ee993b064b37749bc85e928d7f

This temporarily breaks log2 for GCC in cases where GLM_FORCE_PURE
is not defined. The workaround introduced in commit
1ed0e3865b5de0ee993b064b37749bc85e928d7f seems to rely
on getting invalid results from the nlz function.

Broken nlz is caused by a broken findMSB function for GCC.

A fix for the findMSB function should be available in a nearby
separate commit.