Convert integers to type T in quaternion mat3_cast()
This fixes issues when the type T is not explicitly convertible to T, and for example only has an operator* defined that takes two Ts but not a T and an int.

All the other functions in that file already use this syntax, so the missing ones here are most likely an oversight.